# Conflicts:
#	evolution-common-build/pom.xml
#	evolution/pom.xml
#	evolution/src/main/java/com/evolving/nglm/core/DeploymentCommon.java
#	evolution/src/main/java/com/evolving/nglm/evolution/EvolutionEngine.java
#	evolution/src/main/java/com/evolving/nglm/evolution/GUIManager.java
#	evolution/src/main/java/com/evolving/nglm/evolution/Offer.java
#	evolution/src/main/java/com/evolving/nglm/evolution/SubscriberProfile.java
#	evolution/src/main/java/com/evolving/nglm/evolution/SubscriberProfileESSinkConnector.java
#	evolution/src/main/java/com/evolving/nglm/evolution/ThirdPartyJSONGenerator.java
#	evolution/src/main/java/com/evolving/nglm/evolution/ThirdPartyManager.java
#	evolution/src/main/resources/config/deployment-product-evolution.json
#	pom.xml

+++ evolution/src/main/java/com/evolving/nglm/evolution/GUIManager.java
@@ -14636,7 +14636,7 @@ public class GUIManager
                 // incompleteObject
                 //
 
-                IncompleteObject incompleteObject = new IncompleteObject(jsonRoot, epoch, tenantID);
+                IncompleteObject incompleteObject = new IncompleteObject(elementRoot, epoch, tenantID);
 
                 //
                 // store
@@ -25153,7 +25153,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
          *
          *****************************************/
 
-        List<JSONObject> offersJson = offers.stream().map(offer -> ThirdPartyJSONGenerator.generateOfferJSONForThirdParty(offer, offerService, offerObjectiveService, productService, voucherService, salesChannelService, catalogCharacteristicService,false,false
+        List<JSONObject> offersJson = offers.stream().map(offer -> ThirdPartyJSONGenerator.generateOfferJSONForThirdParty(offer, offerService, offerObjectiveService, productService, voucherService, salesChannelService, catalogCharacteristicService, false, false
             )).collect(Collectors.toList());
         response.put("offers", JSONUtilities.encodeArray(offersJson));
         response.put("responseCode", "ok");
@@ -25752,15 +25752,15 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
             Map<String, Object> OfferProductVoucherAndSupplierIDs = OfferProductVoucherAndSupplierIDs(
                 (Offer) existingOffer, tenantID);
             existingSupplierID = (String) OfferProductVoucherAndSupplierIDs.get("supplierID");
-            OfferProduct product = (OfferProduct) OfferProductVoucherAndSupplierIDs.get("offerProduct");
-            OfferVoucher voucher = (OfferVoucher) OfferProductVoucherAndSupplierIDs.get("offerVoucher");
-            if (product != null)
+            JSONObject product = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerProduct");
+            JSONObject voucher = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerVoucher");
+            if (product != null &&  product.get("productID") != null)
               {
-                existingproductID = product.getProductID();
+                existingproductID = product.get("productID").toString();
               }
-            if (voucher != null)
+            if (voucher != null &&  voucher.get("voucherID") != null)
               {
-                existingVoucherID = voucher.getVoucherID();
+                existingVoucherID =  voucher.get("voucherID").toString();
               }
           }
       }
@@ -26109,7 +26109,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
         return JSONUtilities.encodeObject(response);
       }
 
-    else if (activeSupplier != null && activeSupplier != "InactiveReseller")
+    else if (activeSupplier != null && !(activeSupplier.equals("InactiveReseller")))
       {
         for (GUIManagedObject offerObject : offerService.getStoredOffers(tenantID))
           {
@@ -26117,15 +26117,14 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
               {
                 Map<String, Object> OfferProductVoucherAndSupplierIDs = OfferProductVoucherAndSupplierIDs(
                     (Offer) offerObject, tenantID);
-                String supplierID = (String) OfferProductVoucherAndSupplierIDs.get("supplierID");
-                OfferProduct product = (OfferProduct) OfferProductVoucherAndSupplierIDs.get("offerProduct");
-                OfferVoucher voucher = (OfferVoucher) OfferProductVoucherAndSupplierIDs.get("offerVoucher");                
-                
-                Offer offer = (Offer) offerObject;
-                String offerName = offer.getGUIManagedObjectName();                               
-                if (product != null)
+                String supplierID = (String) OfferProductVoucherAndSupplierIDs.get("supplierID"); 
+                JSONObject product = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerProduct");
+                JSONObject voucher = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerVoucher");
+                String offerName = offerObject.getGUIManagedObjectName();
+                Offer offer = (Offer) offerObject;                              
+                if (product != null && product.get("productID") != null)
                   {
-                    String productID = product.getProductID();
+                    String productID = product.get("productID").toString();
                     String productName = (productService.getStoredProduct(productID)).getGUIManagedObjectName();
 
                     if (activeSupplier.equals(supplierID) && offerName.equals(productName) && offer.getSimpleOffer())
@@ -26146,14 +26145,14 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
                           log.debug(offer + " is not supplierOffer");
                       }
                   }
-                if (voucher != null)
+                if (voucher != null && voucher.get("voucherID") != null)
                   {
-                    String voucherID = voucher.getVoucherID();
+                    String voucherID = voucher.get("voucherID").toString();
                     String voucherName = (voucherService.getStoredVoucher(voucherID)).getGUIManagedObjectName();
 
                     if (activeSupplier.equals(supplierID) && offerName.equals(voucherName) && offer.getSimpleOffer())
                       {
-                        vouchers.add(voucher);
+                        //vouchers.add(voucher);
                         offers.add(offer);
                       }
                     else if (offerName.equals(voucherName) && offer.getSimpleOffer() && !(activeSupplier.equals(supplierID)))
@@ -26230,7 +26229,7 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
     String user = JSONUtilities.decodeString(jsonRoot, "loginID", false);
     String activeSupplier = activeSupplierAndParentSupplierIDs(user, tenantID).get("activeSupplierID");
     Date now = SystemTime.getCurrentTime();
-    Offer offer = null;
+    GUIManagedObject offer = null;
 
     /*****************************************
      *
@@ -26240,14 +26239,14 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
     if (jsonRoot.containsKey("id") && offerID != null)
       {
         GUIManagedObject offerObject = offerService.getStoredOffer(offerID);
-        offer = (Offer) offerObject;
+        offer = offerObject;
       }
     else if (jsonRoot.containsKey("offerName") && offerDisplay != null)
       {
         Collection<GUIManagedObject> offers = offerService.getStoredOffers(tenantID);
         for (GUIManagedObject offerObject : offers)
           {
-            Offer currentOffer = (Offer) offerObject;
+            GUIManagedObject currentOffer = offerObject;
             if (currentOffer.getGUIManagedObjectDisplay().equals(offerDisplay))
               {
                 offer = currentOffer;
@@ -26274,30 +26273,35 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
             return JSONUtilities.encodeObject(response);
           }
 
-        else if (activeSupplier != null && activeSupplier != "InactiveReseller")
+        else if (activeSupplier != null && !(activeSupplier.equals("InactiveReseller")))
           {
+            Boolean simpleOffer = false;
+            if (offer.getJSONRepresentation().get("simpleOffer") != null) {
+              simpleOffer = (Boolean) offer.getJSONRepresentation().get("simpleOffer");
+            }
             String offerName = offer.getGUIManagedObjectName();
             Map<String, Object> OfferProductVoucherAndSupplierIDs = OfferProductVoucherAndSupplierIDs(offer, tenantID);
             String supplierID = (String) OfferProductVoucherAndSupplierIDs.get("supplierID");
-            OfferProduct product = (OfferProduct) OfferProductVoucherAndSupplierIDs.get("offerProduct");
-            OfferVoucher voucher = (OfferVoucher) OfferProductVoucherAndSupplierIDs.get("offerVoucher");                   
+            JSONObject product = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerProduct");
+            JSONObject voucher = (JSONObject) OfferProductVoucherAndSupplierIDs.get("offerVoucher");                   
             
-            if (product != null)
+            if (product != null && product.get("productID") != null)
               {
-                String productID = product.getProductID();
+                String productID = product.get("productID").toString();
                 String productName = (productService.getStoredProduct(productID)).getGUIManagedObjectName();
-                if (activeSupplier.equals(supplierID) && offerName.equals(productName) && offer.getSimpleOffer() )
+                
+                if (activeSupplier.equals(supplierID) && offerName.equals(productName) && simpleOffer)
                   {
-                    String productId = product.getProductID();
+                    String productId = productID;
                     productService.removeProduct(productId, userID, tenantID);
-                    offerService.removeOffer(offer.getOfferID(), userID, tenantID);
+                    offerService.removeOffer(offer.getGUIManagedObjectID(), userID, tenantID);
 
                     responseCode = "ok";
 
                     response.put("responseCode", responseCode);
                   }
 
-                else if (offerName.equals(productName) && offer.getSimpleOffer() && !(activeSupplier.equals(supplierID)))
+                else if (offerName.equals(productName) && simpleOffer && !(activeSupplier.equals(supplierID)))
 
                   {
 
@@ -26315,23 +26319,23 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
                     return JSONUtilities.encodeObject(response);
                   }
               }
-            if (voucher != null)
+            if (voucher != null && voucher.get("voucherID") != null)
               {
-                String voucherID = voucher.getVoucherID();
+                String voucherID = voucher.get("voucherID").toString();
                 String voucherName = (voucherService.getStoredVoucher(voucherID)).getGUIManagedObjectName();
 
-                if (activeSupplier.equals(supplierID) && offerName.equals(voucherName) && offer.getSimpleOffer())
+                if (activeSupplier.equals(supplierID) && offerName.equals(voucherName) && simpleOffer)
                   {
-                    String voucherId = voucher.getVoucherID();
+                    String voucherId = voucherID;
                     voucherService.removeVoucher(voucherId, userID, uploadedFileService, tenantID);
-                    offerService.removeOffer(offer.getOfferID(), userID, tenantID);
+                    offerService.removeOffer(offer.getGUIManagedObjectID(), userID, tenantID);
 
                     responseCode = "ok";
 
                     response.put("responseCode", responseCode);
                   }
 
-                else if (offerName.equals(voucherName) && offer.getSimpleOffer() && !(activeSupplier.equals(supplierID)))
+                else if (offerName.equals(voucherName) && simpleOffer && !(activeSupplier.equals(supplierID)))
 
                   {
 
@@ -30101,7 +30105,6 @@ private JSONObject processGetOffersList(String userID, JSONObject jsonRoot, int 
     if (offer != null && offer.getJSONRepresentation().get("vouchers") != null) {
       offerVouchers = (JSONArray) offer.getJSONRepresentation().get("vouchers");
     }
-    
     HashMap<String,Object> response = new HashMap<String,Object>();
     String supplierID = null;
     if (offerProducts != null && offerProducts.size() != 0)
